ai-service/requirements.txt
langgraph==0.0.20
langchain==0.1.6
langchain-openai==0.0.5
openai==1.12.0
httpx==0.26.0
pydantic==2.5.3
python-dotenv==1.0.0

